feat: uniswap compatibility for pair name/symbol
Because Solidly fetches the name and symbol for
the tokens in the pair within the constructor,
token contracts that try to create a pair for the
token itself within its constructor will fail.
Moving these functions outside will ensure that
they will be called only after the token is
created, making ETH/BSC ERC20 contracts
compatible.