feat: make tx revert in LM phase if call reverted
As a safety measure, the add/remove LP functions
will fail if the token supports the Liquidity
Manageable protocol, the router has been appointed
as liquidity manager, but the implementation in
the token contract is faulty and the start/stop LM
phase calls would fail. Previously, they would
silently fail possibly leading to an inconsistent
state. The functions will still silently fail if
the protocol is not implemented in the token, as
it poses no harm.